Add goroutines to increase number of endpoint calls
The current implementation of the code uses a stack to retrieve the data
from the different API endpoints, with a maximum amount of time of 500ms
the amount of data that can be retrieved is significantly lower than
expected. This commit adds parallelism to increase the API calls by
setting the same timeout for all HTTP requests.